CodeLens for #insert section
See https://github.com/redhat-developer/quarkus-ls/issues/814
